replace temp with tmp

+++ src/core/Extension.ts
@@ -4,7 +4,7 @@ import * as fs from "fs-extra";
 import * as https from "https";
 import * as HttpsProxyAgent from "https-proxy-agent";
 import * as path from "path";
-import * as temp from "temp";
+import * as tmp from "tmp";
 import * as vscode from "vscode";
 
 import { IConfig } from "./Config";
@@ -12,7 +12,7 @@ import Environment from "./Environment";
 import Syncing from "./Syncing";
 import * as Toast from "./Toast";
 
-temp.track();
+tmp.setGracefulCleanup();
 
 /**
  * Represent a VSCode extension.
@@ -228,34 +228,42 @@ export default class Extension
     {
         return new Promise((resolve, reject) =>
         {
-            const filepath = temp.path({ suffix: `.${extension.id}.zip` });
-            const file = fs.createWriteStream(filepath);
-            file.on("finish", () =>
+            // Create a temporary file, the file will be automatically closed and unlinked on process exit.
+            tmp.file({ postfix: `.${extension.id}.zip` }, (err, filepath: string) =>
             {
-                resolve(Object.assign({}, extension, { zip: filepath }));
-            }).on("error", reject);
-
-            const options: https.RequestOptions = {
-                host: `${extension.publisher}.gallery.vsassets.io`,
-                path: `/_apis/public/gallery/publisher/${extension.publisher}/extension/${extension.name}/${extension.version}/assetbyname/Microsoft.VisualStudio.Services.VSIXPackage`
-            };
-            const proxy = this._syncing.proxy;
-            if (proxy)
-            {
-                options.agent = new HttpsProxyAgent(proxy);
-            }
-
-            https.get(options, (res) =>
-            {
-                if (res.statusCode === 200)
+                if (err)
                 {
-                    res.pipe(file);
+                    reject(err);
+                    return;
                 }
-                else
+
+                const file = fs.createWriteStream(filepath);
+                file.on("finish", () =>
                 {
-                    reject();
+                    resolve({ ...extension, zip: filepath });
+                }).on("error", reject);
+
+                const options: https.RequestOptions = {
+                    host: `${extension.publisher}.gallery.vsassets.io`,
+                    path: `/_apis/public/gallery/publisher/${extension.publisher}/extension/${extension.name}/${extension.version}/assetbyname/Microsoft.VisualStudio.Services.VSIXPackage`
+                };
+                const proxy = this._syncing.proxy;
+                if (proxy)
+                {
+                    options.agent = new HttpsProxyAgent(proxy);
                 }
-            }).on("error", reject);
+                https.get(options, (res) =>
+                {
+                    if (res.statusCode === 200)
+                    {
+                        res.pipe(file);
+                    }
+                    else
+                    {
+                        reject();
+                    }
+                }).on("error", reject);
+            });
         });
     }
 
@@ -266,46 +274,48 @@ export default class Extension
     {
         return new Promise((resolve, reject) =>
         {
-            temp.mkdir("syncing-", (err1, dirPath) =>
+            const zipFilepath = extension.zip;
+            if (zipFilepath)
             {
-                if (err1)
+                tmp.dir({ postfix: `.${extension.id}`, unsafeCleanup: true }, (err1, dirPath: string) =>
                 {
-                    reject(`Cannot extract extension: ${extension.id}. Access temp folder denied.`);
-                }
-                else if (!extension.zip)
-                {
-                    reject(`Cannot extract extension: ${extension.id}. Zip file not found.`);
-                }
-                else
-                {
-                    extractZip(extension.zip, { dir: dirPath }, (err2) =>
+                    if (err1)
                     {
-                        if (err2)
+                        reject(`Cannot extract extension: ${extension.id}. Access temporary directory denied.`);
+                    }
+                    else
+                    {
+                        extractZip(zipFilepath, { dir: dirPath }, (err2) =>
                         {
-                            reject(`Cannot extract extension: ${extension.id}. ${err2.message}`);
-                        }
-                        else
-                        {
-                            const extPath = path.join(this._env.extensionsPath, `${extension.publisher}.${extension.name}-${extension.version}`);
-                            fs.emptyDir(extPath)
-                                .then(() =>
-                                {
-                                    return fs.copy(path.join(dirPath, "extension"), extPath);
-                                })
-                                .then(() =>
-                                {
-                                    // Clear temp file (background and don't wait).
-                                    fs.remove(extension.zip!).then().catch();
-                                    resolve(Object.assign({}, extension, { path: extPath }));
-                                })
-                                .catch((err3) =>
-                                {
-                                    reject(`Cannot extract extension: ${extension.id}. ${err3.message}`);
-                                });
-                        }
-                    });
-                }
-            });
+                            if (err2)
+                            {
+                                reject(`Cannot extract extension: ${extension.id}. ${err2.message}`);
+                            }
+                            else
+                            {
+                                const extPath = path.join(this._env.extensionsPath, `${extension.publisher}.${extension.name}-${extension.version}`);
+                                fs.emptyDir(extPath)
+                                    .then(() =>
+                                    {
+                                        return fs.copy(path.join(dirPath, "extension"), extPath);
+                                    })
+                                    .then(() =>
+                                    {
+                                        resolve({ ...extension, path: extPath });
+                                    })
+                                    .catch((err3) =>
+                                    {
+                                        reject(`Cannot extract extension: ${extension.id}. ${err3.message}`);
+                                    });
+                            }
+                        });
+                    }
+                });
+            }
+            else
+            {
+                reject(`Cannot extract extension: ${extension.id}. Extension zip file not found.`);
+            }
         });
     }
